added partial admin security

+++ team5caps/src/main/java/sg/iss/team5/controller/AdminHomePageController.java
@@ -28,28 +28,28 @@ import sg.iss.team5.service.AdminService;
 public class AdminHomePageController {
 	@Autowired
 	AdminService adminService;
-	
+
 	Logger logger = LoggerFactory.getLogger(AdminHomePageController.class);
 
 	@RequestMapping(value = "/homepage")
 	public ModelAndView listStudentsNotEnrolled(HttpSession session) {
-		//Security
+		// Security
 		if (!SecurityConfigurations.CheckAdminAuth(session))
 			return new ModelAndView("redirect:/home/login");
-		//Security
-		//get top 7 courses
+		// Security
+		// get top 7 courses
 //		ArrayList<ChartData> chartDataList=adminService.findChartData();
 //		int sum=0;
 //		for (ChartData chartData : chartDataList) {
 //			sum+=chartData.getY();
 //		}
-		
+
 //		//get the average and put into string and int list
 //		String name =chartDataList.get(1).getLabel();
 //		for (int i=0;i<chartDataList.size();i++) {
 //			chartDataList.get(i).setY((int)(chartDataList.get(i).getY()/sum*100));
 //		}
-		
+
 		ArrayList<Student> studentNotEnrolledList = adminService.findNotEnrolled();
 		ModelAndView mav = new ModelAndView("admin_homepage");
 		mav.addObject("listNotEnrolled", studentNotEnrolledList);
@@ -59,8 +59,10 @@ public class AdminHomePageController {
 
 	@RequestMapping(value = "/manage/courses")
 	public ModelAndView listAllCourses(HttpSession session) {
+		// Security
 		if (!SecurityConfigurations.CheckAdminAuth(session))
 			return new ModelAndView("redirect:/home/login");
+		// Security
 		ModelAndView mav = new ModelAndView("admin_courselist");
 		ArrayList<Coursedetail> listAllCourse = adminService.getCourseDetailList();
 		mav.addObject("listAllCourse", listAllCourse);
@@ -73,51 +75,73 @@ public class AdminHomePageController {
 		mav.addObject("enrolledCapacity", capacity);
 		return mav;
 	}
-	
+
 	@RequestMapping(value = "/manage/approval")
-	public ModelAndView approveStudent() {
+	public ModelAndView approveStudent(HttpSession session) {
+		// Security
+		if (!SecurityConfigurations.CheckAdminAuth(session))
+			return new ModelAndView("redirect:/home/login");
+		// Security
 		ArrayList<Studentcourse> sclist = adminService.findByEnrollStatus("Pending");
 		ModelAndView mav = new ModelAndView("admin_approve");
-		mav.addObject("studentcourse",sclist);
+		mav.addObject("studentcourse", sclist);
 		return mav;
 	}
-	
+
 	@RequestMapping(value = "/manage/approval/{mid}/{sid}", method = RequestMethod.POST)
-	public String changeStudentStatus(@PathVariable String mid, @PathVariable String sid, @RequestParam String act) {
+	public String changeStudentStatus(@PathVariable String mid, @PathVariable String sid, @RequestParam String act,
+			HttpSession session) {
+		// Security
+		if (!SecurityConfigurations.CheckAdminAuth(session))
+			return "redirect:/home/login";
+		// Security
+
 		Studentcourse sc = adminService.findByModuleIDCourseID(mid, sid);
 		if (act.equals("Approve")) {
 			sc.setEnrollStatus("Enrolled");
-		}
-		else if (act.equals("Reject")) {
+		} else if (act.equals("Reject")) {
 			sc.setEnrollStatus("Reject");
 		}
 		System.out.println(act);
 		System.out.println(sc);
 		adminService.save(sc);
-		
+
 		return "admin_successful";
 	}
-	
+
 	@RequestMapping(value = "/manage/courses/{cid}")
-	public ModelAndView listStudentsInCourse(@PathVariable String cid) {
+	public ModelAndView listStudentsInCourse(@PathVariable String cid, HttpSession session) {
+		// Security
+		if (!SecurityConfigurations.CheckAdminAuth(session))
+			return new ModelAndView("redirect:/home/login");
+		// Security
+
 		ArrayList<Studentcourse> sclist = adminService.findCourseByCourseId(cid);
 		ModelAndView mav = new ModelAndView("admin_coursedetails");
-		mav.addObject("studentcourse",sclist);
+		mav.addObject("studentcourse", sclist);
 		return mav;
 	}
-	
+
 	@RequestMapping(value = "/manage/courses/{mid}/{sid}", method = RequestMethod.GET)
-	public ModelAndView dropStudent(@PathVariable String mid, @PathVariable String sid) {
+	public ModelAndView dropStudent(@PathVariable String mid, @PathVariable String sid, HttpSession session) {
+		// Security
+		if (!SecurityConfigurations.CheckAdminAuth(session))
+			return new ModelAndView("redirect:/home/login");
+		// Security
 		Student student = adminService.findStudentById(sid);
 		Module module = adminService.findByModuleID(mid);
 		ModelAndView mav = new ModelAndView("admin_dropstudent");
-		mav.addObject("student",student);
-		mav.addObject("module",module);
+		mav.addObject("student", student);
+		mav.addObject("module", module);
 		return mav;
 	}
-	
+
 	@RequestMapping(value = "/manage/courses/{mid}/{sid}", method = RequestMethod.POST)
-	public String dropStudentConfirm(@PathVariable String mid, @PathVariable String sid) {
+	public String dropStudentConfirm(@PathVariable String mid, @PathVariable String sid, HttpSession session) {
+		// Security
+		if (!SecurityConfigurations.CheckAdminAuth(session))
+			return "redirect:/home/login";
+		// Security
 		Studentcourse sc = adminService.findByModuleIDCourseID(mid, sid);
 		adminService.removeStudentCourse(sc);
 		return "admin_successful";
